GraphNode now has edge traversal method
