+++ src/main/scala/com/atomist/tree/pathexpression/NamedNodeTest.scala
@@ -2,13 +2,11 @@ package com.atomist.tree.pathexpression
 
 import com.atomist.graph.GraphNode
 import com.atomist.rug.spi.TypeRegistry
-import com.atomist.tree.TreeNode
-import com.atomist.tree.content.text.TreeNodeOperations
 import com.atomist.tree.pathexpression.ExecutionResult.ExecutionResult
 
 /**
   * Follow children of the given name. If not, methods of that name.
-  * @param name
+  * @param name name of the node or relationship to try
   */
 case class NamedNodeTest(name: String)
   extends NodeTest {
@@ -16,15 +14,7 @@ case class NamedNodeTest(name: String)
   private def findUnder(tn: GraphNode): List[GraphNode] = {
     tn.relatedNodesNamed(name).toList match {
       case Nil =>
-        TreeNodeOperations.invokeMethodIfPresent[TreeNode](tn, name).
-          map {
-            case s: List[TreeNode@unchecked] =>
-              s
-            case t: TreeNode =>
-              List(t)
-            case x =>
-              Nil
-          }.getOrElse(Nil)
+        tn.followEdge(name).toList
       case l => l
     }
   }
